+++ src/test/scala/org/fs/checker/provider/impl/BookTrackerSpec.scala
@@ -13,18 +13,18 @@ import org.scalatest.junit.JUnitRunner
 import com.github.nscala_time.time.Imports._
 
 @RunWith(classOf[JUnitRunner])
-class RutorSpec
+class BookTrackerSpec
     extends FlatSpec
     with TestHelper {
 
-  val instance: Rutor = new Rutor(null, null)
+  val instance: BookTracker = new BookTracker(null, null)
 
-  behavior of "rutor provider"
+  behavior of "booktracker provider"
 
-  it should "parse 2019-02-28 state" in {
-    val content = Source.fromFile(new File(pagesFolder, "young-sheldon_2019-02-28.html"), "UTF-8").mkString
+  it should "parse 08 Jul 2018" in {
+    val content = Source.fromFile(new File(pagesFolder, "on-basilisk-station_2018-07-08.htm"), "UTF-8").mkString
     val parsed = instance.parseDateLastUpdated(content)
-    assert(parsed === DateTime.parse("2019-02-28T23:19:22"))
+    assert(parsed === DateTime.parse("2018-07-08T07:10:00"))
   }
 
   val pagesFolder: java.io.File = new File(resourcesFolder, instance.providerKey)