+++ src/test/scala/org/fs/checker/provider/impl/NonameClubSpec.scala
@@ -13,18 +13,18 @@ import org.scalatest.junit.JUnitRunner
 import com.github.nscala_time.time.Imports._
 
 @RunWith(classOf[JUnitRunner])
-class RutorSpec
+class NonameClubSpec
     extends FlatSpec
     with TestHelper {
 
-  val instance: Rutor = new Rutor(null, null)
+  val instance: NonameClub = new NonameClub(null, null)
 
-  behavior of "rutor provider"
+  behavior of "nnmclub provider"
 
-  it should "parse 2019-02-28 state" in {
-    val content = Source.fromFile(new File(pagesFolder, "young-sheldon_2019-02-28.html"), "UTF-8").mkString
+  it should "parse 09 Jun 2019" in {
+    val content = Source.fromFile(new File(pagesFolder, "elementary_2019-06-09.html"), "windows-1251").mkString
     val parsed = instance.parseDateLastUpdated(content)
-    assert(parsed === DateTime.parse("2019-02-28T23:19:22"))
+    assert(parsed === DateTime.parse("2019-06-09T14:11:53"))
   }
 
   val pagesFolder: java.io.File = new File(resourcesFolder, instance.providerKey)